Moving each project to separate folder Some script cleaning, added helper.py to accomodate the new structure
