+++ utils.py
@@ -10,11 +10,11 @@ DONWLOADS_FOLDER = "/home/olaf/Downloads"
 
 def data_extractor(file_name):
     # Extract the file
-    if not os.path.exists(os.path.join(os.getcwd(), "datasets")):
-        os.makedirs(os.path.join(os.getcwd(), "datasets"))
-    if not os.path.exists(os.path.join(os.path.join(os.getcwd(), f"datasets/{file_name}"))):
+    if not os.path.exists(os.path.join("../datasets")):
+        os.makedirs(os.path.join(os.getcwd(), "../datasets"))
+    if not os.path.exists(os.path.join(os.path.join(f"../datasets/{file_name}"))):
         with zipfile.ZipFile(f"{DONWLOADS_FOLDER}/{file_name}.zip", "r") as zp:
-            zp.extractall(os.path.join(os.getcwd(), f"datasets/{file_name}"))
+            zp.extractall(os.path.join(f"../datasets/{file_name}"))
 
 
 def plot_history(history, fine_tune_epoch=None, title=""):
@@ -38,77 +38,6 @@ def plot_history(history, fine_tune_epoch=None, title=""):
         ax[1].plot([fine_tune_epoch, fine_tune_epoch], [min(acc+val_acc)-0.1, max(acc+val_acc)], label="Fine tune point")
     ax[1].legend(loc="lower right")
     plt.show()
-
-
-def road_classification_model(shape, base_model, augmentation=None):
-    """
-    Returned a model using as feature extractor the base_model
-    Args:
-        shape: tuple, the input shape of the model
-        base_model: the base model
-        augmentation: data augmentation step or steps wrapped in a tf.keras.models.Sequential
-    Returns:
-        tf.keras.Model
-    """
-    base = tf.keras.models.clone_model(base_model)
-    base.set_weights(base_model.get_weights())
-
-    inputs = tf.keras.Input(shape=shape)
-    x = tf.keras.applications.mobilenet_v2.preprocess_input(inputs)
-    if augmentation:
-        x = augmentation(x)
-    x = base(x, training=False)
-    x = tfl.Conv2D(256, (3,3), padding="same")(x)
-    x = tfl.BatchNormalization(axis=1)(x)
-    x = tfl.Activation("relu")(x)
-    x = tfl.Conv2D(512, (3,3))(x)
-    x = tfl.BatchNormalization(axis=1)(x)
-    x = tfl.Activation("relu")(x)
-    x = tfl.MaxPooling2D((2,2))(x)
-    x = tfl.GlobalAveragePooling2D()(x)
-    x = tfl.Dense(400, activation="relu")(x)
-    x = tfl.Dropout(0.2)(x)
-    outputs = tfl.Dense(1)(x)
-
-    return  tf.keras.Model(inputs, outputs)
-
-
-def pollution_model_transfer(shape):
-
-    base_model = tf.keras.applications.mobilenet_v2.MobileNetV2(weights="imagenet", include_top=False, input_shape=shape)
-    base_model.trainable = False
-    inputs = tf.keras.Input(shape=shape)
-    x = tf.keras.applications.mobilenet_v2.preprocess_input(inputs)
-    x = base_model(x , training=False)
-    x = conv_block(x, [256, 256, 512])
-    x = identity_block(x, [256, 256])
-    x = identity_block(x, [512, 512])
-    x = tfl.GlobalAveragePooling2D()(x)
-    x = tfl.Dense(256)(x)
-    x = tfl.Dropout(0.4)(x)
-    x = tfl.Dense(128)(x)
-    x = tfl.Dropout(0.4)(x)
-    outputs = tfl.Dense(11)(x)
-
-    return tf.keras.models.Model(inputs, outputs)
-
-
-def my_pollution_resnet(shape):
-
-    inputs = tf.keras.Input(shape=shape)
-    x = tfl.Conv2D(64, (7,7))(inputs)
-    x = tfl.BatchNormalization(axis=3)(x)
-    x = tfl.Activation("relu")(x)
-    x = tfl.MaxPooling2D((2,2))(x)
-
-    x = conv_block(x, [128,128, 256])
-    x = identity_block(x, [128,128,256])
-    x = identity_block(x, [128,128,256])
-
-    x = tfl.GlobalAveragePooling2D()(x)
-    outputs = tfl.Dense(11)(x)
-
-    return tf.keras.Model(inputs, outputs)
 
 
 def identity_block(input_tensor, filters, kernel_size=(3,3)):
